Extract BaseEntityWithFileManager logic in Trait
